@petermattis 2015-11-02
Does this improve or hurt any of the benchmarks (i.e. the ones in `storage/engine` and `sql`)?