@tamird 2015-11-02
Good question. The results here are all over the place:
```
go test -run=NONE -bench='MVCC|ClientScan|WriteCmd|TxnWrites|Bank|Select' ./... > old.txt
<switch c-rocksdb branches>
go test -run=NONE -bench='MVCC|ClientScan|WriteCmd|TxnWrites|Bank|Select' ./... > new.txt
benchcmp old.txt new.txt
benchmark                                         old ns/op     new ns/op     delta
BenchmarkRPCSSLClientScan1Version1Row-4           536482        601363        +12.09%
BenchmarkRPCSSLClientScan1Version10Rows-4         1284434       1190835       -7.29%
BenchmarkRPCSSLClientScan1Version100Rows-4        5733495       7314280       +27.57%
BenchmarkRPCSSLClientScan1Version1000Rows-4       60985319      59946272      -1.70%
BenchmarkRPCNoSSLClientScan1Version1Row-4         385470        545811        +41.60%
BenchmarkRPCNoSSLClientScan1Version10Rows-4       630015        654400        +3.87%
BenchmarkRPCNoSSLClientScan1Version100Rows-4      4380313       4350150       -0.69%
BenchmarkRPCNoSSLClientScan1Version1000Rows-4     39607734      36989507      -6.61%
BenchmarkTxnWrites-4                              424203        522892        +23.26%
BenchmarkBank-4                                   943581        919403        -2.56%
BenchmarkSelect1-4                                646948        629383        -2.72%
BenchmarkWriteCmdNoEvents-4                       126042        113543        -9.92%
BenchmarkWriteCmdWithEvents-4                     125016        145417        +16.32%
BenchmarkWriteCmdWithEventsAndConsumer-4          120896        131672        +8.91%
BenchmarkMVCCStats-4                              4718          4539          -3.79%
BenchmarkMVCCScan1Version1Row-4                   11982         13823         +15.36%
BenchmarkMVCCScan1Version10Rows-4                 65007         106535        +63.88%
BenchmarkMVCCScan1Version100Rows-4                698599        802431        +14.86%
BenchmarkMVCCScan1Version1000Rows-4               7054805       6261837       -11.24%
BenchmarkMVCCScan10Versions1Row-4                 51482         13242         -74.28%
BenchmarkMVCCScan10Versions10Rows-4               209904        105137        -49.91%
BenchmarkMVCCScan10Versions100Rows-4              1189860       1023504       -13.98%
BenchmarkMVCCScan10Versions1000Rows-4             8102258       10231626      +26.28%
BenchmarkMVCCScan100Versions1Row-4                118145        156924        +32.82%
BenchmarkMVCCScan100Versions10Rows-4              790368        830978        +5.14%
BenchmarkMVCCScan100Versions100Rows-4             6428012       6825431       +6.18%
BenchmarkMVCCScan100Versions1000Rows-4            61324782      58927466      -3.91%
BenchmarkMVCCGet1Version-4                        10870         9676          -10.98%
BenchmarkMVCCGet10Versions-4                      12888         10924         -15.24%
BenchmarkMVCCGet100Versions-4                     68224         107569        +57.67%
BenchmarkMVCCPut10-4                              5320          6389          +20.09%
BenchmarkMVCCPut100-4                             5473          5348          -2.28%
BenchmarkMVCCPut1000-4                            9037          7819          -13.48%
BenchmarkMVCCPut10000-4                           29702         28532         -3.94%
BenchmarkMVCCBatch1Put10-4                        5816          6030          +3.68%
BenchmarkMVCCBatch100Put10-4                      5273          5169          -1.97%
BenchmarkMVCCBatch10000Put10-4                    6947          6595          -5.07%
BenchmarkMVCCBatch100000Put10-4                   7690          7617          -0.95%
BenchmarkMVCCMergeTimeSeries-4                    8063          8329          +3.30%

benchmark                                         old MB/s     new MB/s     speedup
BenchmarkRPCSSLClientScan1Version1Row-4           1.91         1.70         0.89x
BenchmarkRPCSSLClientScan1Version10Rows-4         7.97         8.60         1.08x
BenchmarkRPCSSLClientScan1Version100Rows-4        17.86        14.00        0.78x
BenchmarkRPCSSLClientScan1Version1000Rows-4       16.79        17.08        1.02x
BenchmarkRPCNoSSLClientScan1Version1Row-4         2.66         1.88         0.71x
BenchmarkRPCNoSSLClientScan1Version10Rows-4       16.25        15.65        0.96x
BenchmarkRPCNoSSLClientScan1Version100Rows-4      23.38        23.54        1.01x
BenchmarkRPCNoSSLClientScan1Version1000Rows-4     25.85        27.68        1.07x
BenchmarkMVCCStats-4                              22.04        22.91        1.04x
BenchmarkMVCCScan1Version1Row-4                   85.46        74.08        0.87x
BenchmarkMVCCScan1Version10Rows-4                 157.52       96.12        0.61x
BenchmarkMVCCScan1Version100Rows-4                146.58       127.61       0.87x
BenchmarkMVCCScan1Version1000Rows-4               145.15       163.53       1.13x
BenchmarkMVCCScan10Versions1Row-4                 19.89        77.33        3.89x
BenchmarkMVCCScan10Versions10Rows-4               48.78        97.40        2.00x
BenchmarkMVCCScan10Versions100Rows-4              86.06        100.05       1.16x
BenchmarkMVCCScan10Versions1000Rows-4             126.38       100.08       0.79x
BenchmarkMVCCScan100Versions1Row-4                8.67         6.53         0.75x
BenchmarkMVCCScan100Versions10Rows-4              12.96        12.32        0.95x
BenchmarkMVCCScan100Versions100Rows-4             15.93        15.00        0.94x
BenchmarkMVCCScan100Versions1000Rows-4            16.70        17.38        1.04x
BenchmarkMVCCGet1Version-4                        94.20        105.82       1.12x
BenchmarkMVCCGet10Versions-4                      79.45        93.74        1.18x
BenchmarkMVCCGet100Versions-4                     15.01        9.52         0.63x
BenchmarkMVCCPut10-4                              1.88         1.57         0.84x
BenchmarkMVCCPut100-4                             18.27        18.70        1.02x
BenchmarkMVCCPut1000-4                            110.65       127.89       1.16x
BenchmarkMVCCPut10000-4                           336.67       350.48       1.04x
BenchmarkMVCCBatch1Put10-4                        1.72         1.66         0.97x
BenchmarkMVCCBatch100Put10-4                      1.90         1.93         1.02x
BenchmarkMVCCBatch10000Put10-4                    1.44         1.52         1.06x
BenchmarkMVCCBatch100000Put10-4                   1.30         1.31         1.01x
````

Anyone interested in trying this locally?