@tschottdorf 2015-11-03
```bash
$ go test -run=NONE -bench='MVCCScan1Version1Row' -count 5 -benchtime 2s ./storage/engine > old.txt 2>/dev/null
$ (checkout pr #11 in c-rocksdb)
$ (same with new.txt)
$ benchstat old.txt new.txt
name                    old time/op    new time/op    delta
MVCCScan1Version1Row-8    11.6µs ± 4%    12.0µs ± 7%   ~     (p=0.095 n=5+5)

name                    old speed      new speed      delta
MVCCScan1Version1Row-8  88.7MB/s ± 4%  85.5MB/s ± 7%   ~     (p=0.095 n=5+5)