@tschottdorf 2015-11-03
```
21:37 $ benchstat old.txt new.txt
name                 old time/op   new time/op   delta
MVCCGet10Versions-8   8.56µs ±10%   7.50µs ± 6%  -12.47%  (p=0.016 n=5+5)

name                 old speed     new speed     delta
MVCCGet10Versions-8  120MB/s ±10%  137MB/s ± 6%  +13.94%  (p=0.016 n=5+5)
```